feat: remove plots
